+++ Tests/TestData/MultiFileCharacterization/SourceFiles/WindowsAppVb/WinformsDesignerTest.Designer.vb
@@ -24,7 +24,8 @@ Partial Class WinformsDesignerTest
     Private Sub InitializeComponent()
         Me.Button1 = New System.Windows.Forms.Button() ' In C#, need to assign to field (not property), and bind event manually to ensure Winforms designer renders
         Me.CheckBox1 = New System.Windows.Forms.CheckBox()
-        Me.SuspendLayout
+        Me.Button2 = New System.Windows.Forms.Button()
+        Me.SuspendLayout()
         '
         'Button1
         '
@@ -33,23 +34,33 @@ Partial Class WinformsDesignerTest
         Me.Button1.Size = New System.Drawing.Size(75, 23)
         Me.Button1.TabIndex = 0
         Me.Button1.Text = "Button1"
-        Me.Button1.UseVisualStyleBackColor = true
+        Me.Button1.UseVisualStyleBackColor = True
         '
         'CheckBox1
         '
-        Me.CheckBox1.AutoSize = true
+        Me.CheckBox1.AutoSize = True
         Me.CheckBox1.Location = New System.Drawing.Point(89, 28)
         Me.CheckBox1.Name = "CheckBox1"
         Me.CheckBox1.Size = New System.Drawing.Size(81, 17)
         Me.CheckBox1.TabIndex = 1
         Me.CheckBox1.Text = "CheckBox1"
-        Me.CheckBox1.UseVisualStyleBackColor = true
+        Me.CheckBox1.UseVisualStyleBackColor = True
+        '
+        'Button2
+        '
+        Me.Button2.Location = New System.Drawing.Point(95, 110)
+        Me.Button2.Name = "Button2"
+        Me.Button2.Size = New System.Drawing.Size(75, 23)
+        Me.Button2.TabIndex = 2
+        Me.Button2.Text = "Button2"
+        Me.Button2.UseVisualStyleBackColor = True
         '
         'WinformsDesignerTest
         '
-        Me.AutoScaleDimensions = New System.Drawing.SizeF(6!, 13!)
+        Me.AutoScaleDimensions = New System.Drawing.SizeF(6.0!, 13.0!)
         Me.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font
         Me.ClientSize = New System.Drawing.Size(292, 273)
+        Me.Controls.Add(Me.Button2)
         Me.Controls.Add(Me.CheckBox1)
         Me.Controls.Add(Me.Button1)
         Me.Name = "WinformsDesignerTest"
@@ -64,4 +75,5 @@ End Sub
 
     Friend WithEvents Button1 As Button
     Friend WithEvents CheckBox1 As CheckBox
+    Friend WithEvents Button2 As Button
 End Class